- Updated sidebar component to reduce workload - Updated and renamed useBottomNavbarContent to also render content.
