+++ Website/src/components/app/BottomNavbar/BottomNavbar.tsx
@@ -1,35 +1,10 @@
-import IconButton from "@mui/material/IconButton";
 import Stack from "@mui/material/Stack";
-import Badge, { badgeClasses } from "@mui/material/Badge";
-import { alpha, styled, useTheme } from "@mui/material/styles";
+import { alpha, useTheme } from "@mui/material/styles";
 
-import { useBottomNavbarContent } from "./useBottomNavbarContent";
-import { useActiveTab } from "@/hooks/shared/useActiveTab";
-
-const CustomBadge = styled(Badge)`
-  & .${badgeClasses.badge} {
-    top: -0.5rem;
-    right: 0rem;
-  }
-`;
+import { NavigationButtons } from "./NavigationButtons";
 
 export function BottomNavbar() {
   const theme = useTheme();
-  const { bottomNavbarContent } = useBottomNavbarContent();
-  const { getActiveTab } = useActiveTab();
-
-  const content = bottomNavbarContent.map((item) => {
-    const isActive = getActiveTab(item.label);
-
-    return (
-      <IconButton size="large" onClick={item.handleClick} key={item.label}>
-        {isActive ? item.activeIcon : item.icon}
-        {item.notifications && (
-          <CustomBadge badgeContent={item.notifications} color="primary" overlap="circular" />
-        )}
-      </IconButton>
-    );
-  });
 
   return (
     <Stack
@@ -45,7 +20,7 @@ export function BottomNavbar() {
         left: 0,
       }}
     >
-      {content}
+      <NavigationButtons />
     </Stack>
   );
 }
